Review Hibernate 3.6.5 patch and OpenMRS upgrade patch - SEARCH-10 Author: czakian
git-svn-id: http://svn.openmrs.org/openmrs/trunk@24899 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/db/PatientDAO.java
@@ -148,6 +148,6 @@ public interface PatientDAO {
 	/**
 	 * @see PatientService#getCountOfPatients(String)
 	 */
-	public Integer getCountOfPatients(String name, String identifier, List<PatientIdentifierType> identifierTypes,
+	public Long getCountOfPatients(String name, String identifier, List<PatientIdentifierType> identifierTypes,
 	        boolean matchIdentifierExactly);
 }
